requirements.txt
aiofiles
bard
SpeechRecognition
Flask>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
aiohttp 
asyncio 
beautifulsoup4
daxxhub
deepai
python-whois
dnspython
ffmpeg-python
gitpython
hachoir
Faker
pydub
pycountry
heroku3
apscheduler
motor 
numpy
pillow==9.5.0
psutil
wget
pyfiglet
qrcode
pytz
future==0.18.3
search_engine_parser
py-tgcalls==1.2.9
ntgcalls==1.1.2
pykeyboard
pyshorteners
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
pyrofork
python-dotenv
pyyaml
requests
emojis
speedtest-cli
spotipy
tgcrypto
telegraph
unidecode
youtube-search
youtube-search-python
gtts
git+https://github.com/AsmSafone/SafoneAPI
lexica-api==1.5.1
httpx==0.27.2
gpytranslate
openai
pickledb
python-telegram-bot==20.7
img2pdf
cloudscraper
pycryptodome
opencv-python
yt-dlp>=2025.6.25
